Reorg tests to make them PSR-0 compatible
