+++ phpunit.xml
@@ -5,7 +5,7 @@
 >
     <testsuites>
         <testsuite name="Unit Tests">
-            <directory>tests/unit</directory>
+            <directory>tests</directory>
         </testsuite>
     </testsuites>
